modal_app/requirements.txt
modal
boto3
requests
Pillow
fastapi
